Pack modCovColors as ABGR u32 in RPC worker
Same treatment as readTagColors / modificationColors: the worker emits
one u32 per mod-coverage segment; packModCovSegmentsForGpu reads the
u32 directly into its vertex buffer instead of shifting 4 loose bytes.

+++ plugins/alignments/src/LinearAlignmentsDisplay/components/rendererTypes.ts
@@ -123,7 +123,7 @@ export interface ModCoverageUploadData {
   modCovPositions: Uint32Array
   modCovYOffsets: Float32Array
   modCovHeights: Float32Array
-  modCovColors: Uint8Array
+  modCovColors: Uint32Array
   numModCovSegments: number
   modCovPackedBuffer: ArrayBuffer
 }
